requirements.txt
certifi==2018.11.29
chardet==3.0.4
click==7.0
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4
numpy==1.16.1
openpyxl==2.6.0
pandas==0.24.1
python-dateutil==2.8.0
pytz==2018.9
requests==2.21.0
six==1.12.0
urllib3==1.24.1